+++ api/src/application/user/user-controller.js
@@ -12,6 +12,7 @@ const comparePasswords = require('../../lib/comparePasswords');
 const bookshelf = require('../../lib/bookshelf');
 const webhookService = require('../webhook/webhook-service');
 const logger = require('../../lib/logger');
+const allowedImageMimes = require('../image/allowed-image-mimes');
 
 // e.g. convert { foo.bar: 'baz' } to { foo: { bar: 'baz' }}
 const expandDotPaths = function(object) {
@@ -101,10 +102,11 @@ module.exports = (
       const { shouldClearPicture, ...originalPayload } = request.payload;
       const payload = expandDotPaths(originalPayload);
 
+      const uploadingNewPicture = !!originalPayload.picture._data.byteLength;
       const oldPicture = profile.picture;
       const pictureMIME = originalPayload.picture.hapi.headers['content-type'];
 
-      if (pictureMIME === 'image/jpeg' || pictureMIME === 'image/png') {
+      if (uploadingNewPicture && allowedImageMimes.indexOf(pictureMIME) >= 0) {
         const uuid = Uuid();
         const bucket = uuid.substring(0, 2);
         const filename = await imageService.uploadImageStream(originalPayload.picture, `pictures/${bucket}/${uuid}`);
@@ -122,7 +124,7 @@ module.exports = (
       user = await userService.update(user.get('id'), { profile });
       webhookService.trigger('user.update', user);
 
-      if (oldPicture) {
+      if ((uploadingNewPicture && oldPicture) || shouldClearPicture) {
         await imageService.deleteImage(oldPicture.replace(/^.*\/\/[^\/]+\//, ''));
       }
 
@@ -137,12 +139,8 @@ module.exports = (
       }
 
       const viewContext = views.forgotPasswordSuccess(client, request);
-      const template = await themeService.renderThemedTemplate(request.query.client_id, 'forgot-password-success', viewContext);
-      if (template) {
-        reply(template);
-      } else {
-        reply.view('forgot-password-success', viewContext);
-      }
+      const template = await themeService.renderThemedTemplate('forgot-password-success', viewContext, request.query.client_id);
+      reply(template);
     }),
 
     completeEmailUpdateHandler: async (request, reply, source, error) => {
@@ -172,12 +170,8 @@ module.exports = (
 
       const viewContext = views.completeChangePassword(user, client, request, error);
 
-      const template = await themeService.renderThemedTemplate(request.query.client_id, 'email-verify-success', viewContext);
-      if (template) {
-        return reply(template);
-      } else {
-        return reply.view('email-verify-success', viewContext);
-      }
+      const template = await themeService.renderThemedTemplate('email-verify-success', viewContext, request.query.client_id);
+      return reply(template);
     },
 
     emailVerifySuccessHandler: async (request, reply, source, error) => {
@@ -196,12 +190,8 @@ module.exports = (
       const client = await clientService.findById(request.query.client_id);
       const viewContext = views.emailVerifySuccess(user, client, request, error);
 
-      const template = await themeService.renderThemedTemplate(request.query.client_id, 'email-verify-success', viewContext);
-      if (template) {
-        reply(template);
-      } else {
-        reply.view('email-verify-success', viewContext);
-      }
+      const template = await themeService.renderThemedTemplate('email-verify-success', viewContext, request.query.client_id);
+      reply(template);
     },
 
     resetPassword: async function(request, reply, user, client, render) {
@@ -213,12 +203,8 @@ module.exports = (
       await emailTokenService.destroyUserTokens(user.get('id'));
 
       const viewContext = views.resetPasswordSuccess(request);
-      const template = await themeService.renderThemedTemplate(request.query.client_id, 'reset-password-success', viewContext);
-      if (template) {
-        reply(template);
-      } else {
-        reply.view('reset-password-success', viewContext);
-      }
+      const template = await themeService.renderThemedTemplate('reset-password-success', viewContext, request.query.client_id);
+      reply(template);
     },
 
     logout: function(request, reply) {