+++ api/src/plugins/openid-connect/getConfig.js
@@ -58,20 +58,20 @@ module.exports = options => {
     logoutSource: async function renderLogoutSource(ctx, form) {
       const clientId = ctx.oidc.session.logout.clientId;
       const template = await options.getTemplate(clientId, 'end-session', { form });
-      if (template) {
-        ctx.body = template;
-      } else {
-        const layout = fs.readFileSync('./templates/layout/default.hbs', 'utf8');
-        const logout = fs.readFileSync('./templates/end_session.hbs', 'utf8');
-
-        ctx.body = handlebars.compile(layout)({
-          content: handlebars.compile(logout)({ form })
-        });
-      }
+      ctx.body = template;
     },
     subjectTypes: ['public', 'pairwise'],
     pairwiseSalt: options.pairwiseSalt,
     interactionUrl: async (ctx, interaction) => `/interaction/${ctx.oidc.uuid}`,
-    scopes: ['admin', 'superadmin'],
-  }
-}
+    scopes: ['admin', 'superadmin', 'offline_access'],
+    ttl: {
+      AccessToken: options.ttl.AccessToken || 3600,
+      AuthorizationCode: options.ttl.AuthorizationCode || 600,
+      ClientCredentials: options.ttl.ClientCredentials || 600,
+      DeviceCode: options.ttl.DeviceCode || 600,
+      IdToken: options.ttl.IdToken || 3600,
+      RefreshToken: options.ttl.RefreshToken || 1209600 // default two weeks
+    },
+    renderError: options.renderError,
+  };
+};