+++ api/src/plugins/openid-connect/grants/password.js
@@ -6,11 +6,12 @@ module.exports = (options) => ({
       const account = await options.authenticateUser(username, password);
 
       if (account) {
-        const { AccessToken, IdToken } = providerInstance;
+        const { AccessToken, IdToken, RefreshToken } = providerInstance;
         const at = new AccessToken({
-          accountId: 'foo',
+          accountId: account.accountId,
           clientId: ctx.oidc.client.clientId,
           grantId: ctx.oidc.uuid,
+          scope: ctx.oidc.params.scope || '',
         });
 
         const accessToken = await at.save();
@@ -18,9 +19,24 @@ module.exports = (options) => ({
 
         const token = new IdToken(
           Object.assign({}, await Promise.resolve(account.claims())),
-          ctx.oidc.client.sectorIdentifier
+          ctx.oidc.client,
         );
+
+        const refreshToken = new RefreshToken({
+          clientId: ctx.oidc.client.clientId,
+          scope: ctx.oidc.params.scope || '',
+          accountId: account.accountId,
+          grantId: ctx.oidc.uuid,
+          claims: {
+            id_token: { sub: { value: account.accountId } }
+          }
+        });
+
+        const refreshTokenValue = await refreshToken.save();
+
         token.set('at_hash', accessToken);
+        token.set('rt_hash', refreshTokenValue);
+        token.set('sub', account.accountId);
 
         const idToken = await token.sign(ctx.oidc.client);
 
@@ -29,12 +45,14 @@ module.exports = (options) => ({
           expires_in: expiresIn,
           token_type: 'Bearer',
           id_token: idToken,
+          refresh_token: refreshTokenValue,
         };
       } else {
         ctx.body = {
           error: 'invalid_grant',
           error_description: 'invalid credentials provided',
         };
+        ctx.status = 400;
       }
 
       await next();