+++ test-client/src/App.js
@@ -6,6 +6,7 @@ import config from './config';
 import './App.css';
 
 import InviteUserForm from './components/InviteUserForm';
+import UserProfileForm from './components/UserProfileForm';
 
 const history = createHistory();
 
@@ -13,7 +14,28 @@ class App extends Component {
   constructor(props) {
     super(props);
 
-    this.state = {};
+    this.state = {
+      email: '',
+      password: '',
+    };
+  }
+
+  login = () => {
+    fetch('/token', {
+      method: 'POST',
+      body: JSON.stringify({ grant_type: 'password', username: this.state.email, password: this.state.password }),
+    })
+      .then(res => res.json())
+      .then(json => {
+        localstorage({
+          accessToken: json.access_token,
+          expiresIn: json.expires_in,
+          idToken: json.id_token,
+          tokenType: json.token_type,
+        });
+        this.setState(localstorage());
+      })
+      .catch(e => console.error(e));
   }
 
   componentWillMount() {
@@ -50,6 +72,13 @@ class App extends Component {
           <span> | </span>
           <a href={`https://sso-client.test:9000/user/register?client_id=${config.clientId}&response_type=code id_token token&scope=${config.scope}&redirect_uri=${config.redirectUri}&nonce=nonce`}>Sign Up</a>
         </span>
+
+        <div>
+          <input type="text" id="email" value={this.state.email} onChange={({ target }) => this.setState(() => ({ email: target.value }))} />
+          <input type="password" id="password" value={this.state.password} onChange={({ target }) => this.setState(() => ({ password: target.value }))}/>
+          <button onClick={this.login}>login with password grant</button>
+          <a href={`https://sso-client.test:9000/user/forgot-password?client_id=${config.clientId}&redirect_uri=${config.redirectUri}`}>Forgot Password</a>
+        </div>
       </div>
     );
   }
@@ -70,7 +99,8 @@ class App extends Component {
         </div>
         <div>
           <InviteUserForm />
-          <div><a href={`https://sso-client.test:9000/user/profile?client_id=${config.clientId}&redirect_uri=${config.redirectUri}`}>Edit Profile</a></div>
+          <UserProfileForm />
+          <div><a href={`https://sso-client.test:9000/user/profile?client_id=${config.clientId}&redirect_uri=${config.redirectUri}&access_token=${this.state.accessToken}`}>Edit Profile</a></div>
           <div><a href={`https://sso-client.test:9000/user/password?client_id=${config.clientId}&redirect_uri=${config.redirectUri}`}>Change Password</a></div>
           <div><a href={`https://sso-client.test:9000/user/email-settings?client_id=${config.clientId}&redirect_uri=${config.redirectUri}`}>Email Settings</a></div>
           <div><a href={`https://sso-client.test:9000/op/session/end?id_token_hint=${this.state.idToken}&post_logout_redirect_uri=${config.postLogoutRedirectUri}`}>Log Out</a></div>